Wire DynamoAdminManager to AdministrationService, refactor package name for dynamo managers
