+++ services/repository-managers/src/main/java/org/sagebionetworks/repo/manager/dynamo/NodeTreeQueryManager.java
@@ -1,4 +1,4 @@
-package org.sagebionetworks.repo.manager;
+package org.sagebionetworks.repo.manager.dynamo;
 
 import org.sagebionetworks.repo.model.DatastoreException;
 import org.sagebionetworks.repo.model.EntityId;
