+++ services/repository-managers/src/test/java/org/sagebionetworks/repo/manager/dynamo/DynamoAdminManagerImplTest.java
@@ -1,4 +1,4 @@
-package org.sagebionetworks.repo.manager;
+package org.sagebionetworks.repo.manager.dynamo;
 
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
@@ -8,6 +8,8 @@ import static org.mockito.Mockito.when;
 import org.junit.Assert;
 import org.junit.Test;
 import org.sagebionetworks.dynamo.dao.DynamoAdminDao;
+import org.sagebionetworks.repo.manager.UserManager;
+import org.sagebionetworks.repo.manager.dynamo.DynamoAdminManagerImpl;
 import org.sagebionetworks.repo.model.UnauthorizedException;
 import org.sagebionetworks.repo.model.UserInfo;
 import org.springframework.test.util.ReflectionTestUtils;
