+++ services/repository/src/main/java/org/sagebionetworks/repo/web/service/AdministrationService.java
@@ -193,4 +193,9 @@ public interface AdministrationService {
 	 * Clears the Synapse DOI table.
 	 */
 	void clearDoi(String userId) throws NotFoundException, UnauthorizedException, DatastoreException;
+
+	/**
+	 * Clears the specified dynamo table.
+	 */
+	void clearDynamoTable(String userId, String tableName, String hashKeyName, String rangeKeyName) throws NotFoundException, UnauthorizedException, DatastoreException;
 }
