+++ services/repository/src/main/java/org/sagebionetworks/repo/web/service/AdministrationServiceImpl.java
@@ -10,6 +10,7 @@ import org.sagebionetworks.repo.manager.UserManager;
 import org.sagebionetworks.repo.manager.backup.daemon.BackupDaemonLauncher;
 import org.sagebionetworks.repo.manager.backup.migration.DependencyManager;
 import org.sagebionetworks.repo.manager.doi.DoiAdminManager;
+import org.sagebionetworks.repo.manager.dynamo.DynamoAdminManager;
 import org.sagebionetworks.repo.manager.message.MessageSyndication;
 import org.sagebionetworks.repo.model.ConflictingUpdateException;
 import org.sagebionetworks.repo.model.DatastoreException;
@@ -55,6 +56,8 @@ public class AdministrationServiceImpl implements AdministrationService  {
 	MessageSyndication messageSyndication;
 	@Autowired
 	private DoiAdminManager doiAdminManager;
+	@Autowired
+	private DynamoAdminManager dynamoAdminManager;
 
 	/**
 	 * Spring will use this constructor
@@ -247,5 +250,12 @@ public class AdministrationServiceImpl implements AdministrationService  {
 	@Override
 	public void clearDoi(String userId) throws NotFoundException, UnauthorizedException, DatastoreException {
 		doiAdminManager.clear(userId);
+	}
+
+	@Override
+	public void clearDynamoTable(String userId, String tableName,
+			String hashKeyName, String rangeKeyName) throws NotFoundException,
+			UnauthorizedException, DatastoreException {
+		dynamoAdminManager.clear(userId, tableName, hashKeyName, rangeKeyName);
 	}
 }
